Initial version of the iniManager class.
See comments in the class for additional info.  The class itself should be easy enough to follow.  It merely reads an INI file into a multidimensional array so the developer can manipulate it, add new sections easily, etc. and then display the INI file or write it back to the original file.